Added new animations

+++ Assets/Player/SwitchCharacter.cs
@@ -24,7 +24,7 @@ public class SwitchCharacter : MonoBehaviour
 
         GameObject player = GameObject.FindGameObjectWithTag("Player");
         if (player != null) PossessPlayer(player);
-        
+
         cameraManager = GameObject.FindGameObjectWithTag("MainCamera");
     }
 
@@ -62,57 +62,50 @@ public class SwitchCharacter : MonoBehaviour
 
     void CycleToNextPikmin()
     {
-        //get cinemachine target group reference
         CinemachineTargetGroup group = cameraManager.GetComponentInChildren<CinemachineTargetGroup>();
-        
-        // Build list from main player position if we're on the main player
+
         if (_sortedPikmin.Count == 0)
             BuildSortedPikminList();
 
         if (_sortedPikmin.Count == 0) return;
 
+        // Remove any null or destroyed pikmin from the list
+        _sortedPikmin.RemoveAll(p => p == null);
+
+        if (_sortedPikmin.Count == 0) return;
+
         this.gameObject.GetComponent<SignalController>().Deactivate();
-        
+
         _currentPikminIndex = (_currentPikminIndex + 1) % _sortedPikmin.Count;
-        GameObject currentPikmin = _sortedPikmin[_currentPikminIndex].gameObject;
+        GameObject currentPikmin = _sortedPikmin[_currentPikminIndex];
+
+        if (currentPikmin == null) return;
 
         GetComponent<Rigidbody2D>().linearVelocityX = 0;
         PossessPlayer(currentPikmin);
         this.gameObject.GetComponent<SignalController>().LittleDude = currentPikmin;
-    
-        //if player character, activate signal controller
+
         if (gameObject.CompareTag("Player"))
         {
             if (group.Targets.Count > 1)
-            {
                 group.Targets.RemoveRange(1, group.Targets.Count - 1);
-            }
+
             this.gameObject.GetComponent<SignalController>().Activate();
-            
-            //if cameraManager is also valid add the little dude to the target group
-            if(cameraManager)
-            {
+
+            if (cameraManager)
                 group.AddMember(currentPikmin.transform, 1f, 0.5f);
-            }
-            
         }
-       
-       
-       
     }
 
     public void ReturnToPlayer()
     {
         _sortedPikmin.Clear();
         _currentPikminIndex = -1;
-        
-        //get cinemachine target group reference
+
         CinemachineTargetGroup group = cameraManager.GetComponentInChildren<CinemachineTargetGroup>();
         if (group.Targets.Count > 1)
-        {
             group.Targets.RemoveRange(1, group.Targets.Count - 1);
-        }
-        
+
         this.gameObject.GetComponent<SignalController>().Deactivate();
 
         GameObject player = GameObject.FindGameObjectWithTag("Player");
